Update test/run-tests.py
* Write all intermediate output to out/
* Use real paths for all test names (i.e. include "test/")
* A few Python3 fixes

+++ test/gen-spec-js.py
@@ -307,7 +307,7 @@ class ModuleExtender(object):
     elif type_ == 'f32':
       inst = F32ToWasm(int(const['value']))
     elif type_ == 'f64':
-      inst = F64ToWasm(long(const['value']))
+      inst = F64ToWasm(int(const['value']))
     self.lines.append(inst)
 
   def _RunWasm2Wast(self, wasm_path):
